build: update relative import paths on all files
This change is necessary because when trying to run the application,
Node was looking for the exact files in the import path (e.g. ./routes).
However, for Nodes module resolution to work properly, it needs the
exact correct file name (e.g. ./routes.js).
For this purpose, the build script has also been updated to copy the
views folder in the correct place and try to run the server.js from the
correct directoy within /dist.