fix: fix POI API tests
New Joi schemas were introduced to allow validation to accept both
object with a nested location object and without (top level 'lat' and
'lng'). This way when using the API, an object with nested location can
be sent, and when using the user interface with forms (which can only
send flat objects) function also correctly.