test, fix: rename auth api test to run after other API test suites
The reason this is necessary is, that the first API test suite that runs
after the auth-api-test suite always fails in the setup. Fixing this is
something that should be done in the future, however I couldn't get it
to work.

I tried adding a teardown after the tests that clears the authentication
and deletes all users, however this didn't resolve the issue.

In the API tests that follow, it sometimes fails when trying to
authenticate a user in the test setup to clean the database (delete all
users and categories etc.), and sometimes fails after authentication,
even if the correct authorization header gets set, when trying to
execute the cleanup methods.